==> 1-ewas/1-buccal/4-gsea.Rmd <==
---
title: "Enrichment analysis - missMethyl and clusterProfiler"
output: html_document
---

# Settings

```{r}
library(missMethyl)
library(IlluminaHumanMethylationEPICanno.ilm10b4.hg19)
library(clusterProfiler)
library(org.Hs.eg.db)
library(ReactomePA)

# Settings
here::i_am("1-ewas/1-buccal/3-dmr-visualisation.Rmd")
```

# Load DMPs

```{r}
buccal <- readRDS(here("1-ewas", "1-buccal", "delta-beta.Rds")) |> 
  as.data.frame() |> 
  dplyr::mutate(padj = p.adjust(type, method = 'BH')) |> 
  tibble::rownames_to_column('cg')

sigCpGs <- subset(buccal, padj < 0.05, select = cg)
sigCpGs <- dplyr::pull(sigCpGs,cg)

# Write DMPs to .txt file for eFORGE upload
writeLines(sigCpGs, here("1-ewas", "1-buccal", "dmps.txt"))
```


# missMethyl

## Buccal only

```{r}
library(IlluminaHumanMethylation450kanno.ilmn12.hg19)
gst.go <- gometh(sig.cpg=sigCpGs, collection="GO", array.type="EPIC",
              plot.bias=TRUE)

gst.kegg <- gometh(sig.cpg=sigCpGs, collection="KEGG")

save(gst.go, gst.kegg,file=here("1-ewas","1-buccal","2-output","gsea-missMethyl.Rdata"))

load(here("1-ewas","1-buccal","2-output","gsea-missMethyl.Rdata"))
topGSA(gst.go)
topGSA(gst.kegg)
```

## Overlap buccal - cervical

```{r}
cervical <- readRDS(here("1-ewas", "2-cervical", "delta-beta.Rds")) |> 
  as.data.frame() |> 
  dplyr::mutate(padj = p.adjust(type, method = 'BH'))|> 
  tibble::rownames_to_column('cg') 

dmps <- buccal |> 
  dplyr::full_join(cervical, by = 'cg') |> 
  dplyr::filter(padj.y < 0.05 & padj.x < 0.05)


sigCpGs <- dplyr::pull(dmps,cg) # 37

gst.go <- gometh(sig.cpg=sigCpGs, collection="GO", array.type="EPIC",
              plot.bias=TRUE)

gst.kegg <- gometh(sig.cpg=sigCpGs, collection="KEGG")

save(gst.go, gst.kegg,file=here("1-ewas","1-buccal","2-output","gsea-missMethyl-overlap-buccal-cervical.Rdata"))

load(here("1-ewas","1-buccal","2-output","gsea-missMethyl-overlap-buccal-cervical.Rdata"))
topGSA(gst.go)
topGSA(gst.kegg)
```

# ClusterProfiler

```{r}
anno <- getAnnotation(IlluminaHumanMethylationEPICanno.ilm10b4.hg19)
anno$gene  <- stringr::str_split(anno$UCSC_RefGene_Name, ";", simplify = T)[,1]

# Universe - all unique genes
universe <- unique(anno$gene)
universe <- universe[universe != ""] # remove empty (for those CpGs not in genes)
```

## Buccal only

```{r}
buccal <- readRDS(here("1-ewas", "1-buccal", "delta-beta.Rds")) |> 
  as.data.frame() |> 
  dplyr::mutate(padj = p.adjust(type, method = 'BH')) |> 
  tibble::rownames_to_column('cg')

sigCpGs <- subset(buccal, padj < 0.05, select = cg)
sigCpGs <- dplyr::pull(sigCpGs,cg)

# Genes of interest (goi) in our list
goi <- unique(anno[anno$Name %in% sigCpGs,]$gene)

# ENTREZIDS
universegenes <- bitr(goi, fromType="SYMBOL", toType="ENTREZID", OrgDb="org.Hs.eg.db")

# Biological process
bp <- clusterProfiler::enrichGO(gene = goi,
                                universe = universe,
                                OrgDb = org.Hs.eg.db,
                                keyType = 'SYMBOL',
                                ont = "BP",
                                pAdjustMethod = "BH")
#bp <- as.data.frame(bp) # extract as dataframe 

# GO molecular function
mf <- clusterProfiler::enrichGO(gene = goi,
                                universe = universe,
                                OrgDb = org.Hs.eg.db,
                                keyType = 'SYMBOL',
                                ont = "mf",
                                pAdjustMethod = "BH")
#mf <- as.data.frame(mf)

# GO cellular process
cc <- clusterProfiler::enrichGO(gene = goi,
                                universe = universe,
                                OrgDb = org.Hs.eg.db,
                                keyType = 'SYMBOL',
                                ont = "CC",
                                pAdjustMethod = "BH")

#cc <- as.data.frame(cc)

# Reactome pathway enrichment 
goi_entrez <- bitr(goi, fromType="SYMBOL", toType="ENTREZID", OrgDb="org.Hs.eg.db")
universe_entrez <- bitr(universe, fromType="SYMBOL", toType="ENTREZID", OrgDb="org.Hs.eg.db")
epi_hypo_react_all <- enrichPathway(goi_entrez$ENTREZID, organism = "human",
                                    universe = universe_entrez$ENTREZID,
                                    pAdjustMethod = "BH", 
                                    readable = T,
                                    minGSSize = 3)


## save results -------------------
save(bp, mf, cc, epi_hypo_react_all, file=here("1-ewas","1-buccal","2-output","gsea-clusterProfiler.Rdata"))

# Visualize
bp_dot <- enrichplot::dotplot(bp) 
```

Only BP enriched.

## Overlap buccal - cervical

```{r}
cervical <- readRDS(here("1-ewas", "2-cervical", "delta-beta.Rds")) |> 
  as.data.frame() |> 
  dplyr::mutate(padj = p.adjust(type, method = 'BH'))|> 
  tibble::rownames_to_column('cg') 

dmps <- buccal |> 
  dplyr::full_join(cervical, by = 'cg') |> 
  dplyr::filter(padj.y < 0.05 & padj.x < 0.05)


sigCpGs <- dplyr::pull(dmps,cg) # 37

# Genes of interest (goi) in our list
goi <- unique(anno[anno$Name %in% sigCpGs,]$gene)

# ENTREZIDS
universegenes <- bitr(goi, fromType="SYMBOL", toType="ENTREZID", OrgDb="org.Hs.eg.db")

# Biological process
bp <- clusterProfiler::enrichGO(gene = goi,
                                universe = universe,
                                OrgDb = org.Hs.eg.db,
                                keyType = 'SYMBOL',
                                ont = "BP",
                                pAdjustMethod = "BH")
#bp <- as.data.frame(bp) # extract as dataframe 

# GO molecular function
mf <- clusterProfiler::enrichGO(gene = goi,
                                universe = universe,
                                OrgDb = org.Hs.eg.db,
                                keyType = 'SYMBOL',
                                ont = "mf",
                                pAdjustMethod = "BH")
#mf <- as.data.frame(mf)

# GO cellular process
cc <- clusterProfiler::enrichGO(gene = goi,
                                universe = universe,
                                OrgDb = org.Hs.eg.db,
                                keyType = 'SYMBOL',
                                ont = "CC",
                                pAdjustMethod = "BH")

#cc <- as.data.frame(cc)

# Reactome pathway enrichment 
goi_entrez <- bitr(goi, fromType="SYMBOL", toType="ENTREZID", OrgDb="org.Hs.eg.db")
universe_entrez <- bitr(universe, fromType="SYMBOL", toType="ENTREZID", OrgDb="org.Hs.eg.db")
epi_hypo_react_all <- enrichPathway(goi_entrez$ENTREZID, organism = "human",
                                    universe = universe_entrez$ENTREZID,
                                    pAdjustMethod = "BH", 
                                    readable = T,
                                    minGSSize = 3)
```

Does not yield any enrichments.
